+++ src/layouts/CMSLayout.astro
@@ -1,6 +1,6 @@
 ---
 import Aside from "~/components/Aside.astro";
-import { getGithubEditUrl } from "~/util";
+import { getGithubEditUrl } from "~/util/getGithubEditUrl";
 import MainLayout from "./MainLayout.astro";
 import UIString from '~/components/UIString.astro';
