+++ src/layouts/MainLayout.astro
@@ -3,7 +3,7 @@ import PageContent from '../components/PageContent/PageContent.astro';
 import RightSidebar from '../components/RightSidebar/RightSidebar.astro';
 import TableOfContents from '../components/RightSidebar/TableOfContents';
 import { useTranslations } from '../i18n/util';
-import { getGithubEditUrl } from '../util';
+import { getGithubEditUrl } from '../util/getGithubEditUrl';
 import { getNavLinks } from '../util/getNavLinks';
 import BaseLayout from './BaseLayout.astro';
