+++ src/util.ts
@@ -1,4 +1,3 @@
-import { AstroGlobal } from "astro";
 
 export function getLanguageFromURL(pathname: string) {
 	const langCodeMatch = pathname.match(/\/([a-z]{2}-?[a-z]{0,2})\//);
@@ -24,18 +23,4 @@ export function removeSubpageSegment(path: string) {
 	return path;
 }
 
-/** Gets the URL to edit the page on GitHub */
-export function getGithubEditUrl(Astro: Readonly<AstroGlobal>) {
-	const { content = {} } = Astro.props;
-	const isFallback = !!Astro.params.fallback;
-	const currentPage = Astro.url.pathname;
-	const lang = getLanguageFromURL(currentPage);
-	const filePath = `src/pages${currentPage.replace(/\/$/, '')}.md`;
-	const currentFile = isFallback ? filePath.replace(`/${lang}/`, '/en/') : filePath;
-	const githubEditUrl =
-		content.githubURL && (lang === 'en' || isFallback)
-			? `${content.githubURL}${content.hasREADME ? 'README.md' : ''}`
-			: `https://github.com/withastro/docs/blob/main/${currentFile}`;
 
-	return githubEditUrl;
-}
