+++ src/util/getGithubEditUrl.ts
@@ -1,0 +1,18 @@
+import { AstroGlobal } from "astro";
+import { getLanguageFromURL } from "../util";
+
+/** Gets the URL to edit the page on GitHub */
+
+export function getGithubEditUrl(Astro: Readonly<AstroGlobal>) {
+	const { content = {} } = Astro.props;
+	const isFallback = !!Astro.params.fallback;
+	const currentPage = Astro.url.pathname;
+	const lang = getLanguageFromURL(currentPage);
+	const filePath = `src/pages${currentPage.replace(/\/$/, '')}.md`;
+	const currentFile = isFallback ? filePath.replace(`/${lang}/`, '/en/') : filePath;
+	const githubEditUrl = content.githubURL && (lang === 'en' || isFallback)
+		? `${content.githubURL}${content.hasREADME ? 'README.md' : ''}`
+		: `https://github.com/withastro/docs/blob/main/${currentFile}`;
+
+	return githubEditUrl;
+}
